requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
requests
geopy
python-dotenv
pydantic-settings
alembic
